Refactor MeasurementData: mover la interfaz MeasurementData a un lugar centralizado para mejorar la consistencia y la reutilización en los hooks.
